compositor|refactor: Name the plan traversal module for its export
The plan's traversal module is now named for its one export, `buildTraversalIndex`, so it no longer shares a filename with the dependency-graph module it composes.
