+++ packages/compositor/src/plan/buildTraversalIndex.ts
